# Optimize asynchronous merger