--- a/crates/dbsp/src/operator/csv.rs
+++ b/crates/dbsp/src/operator/csv.rs
@@ -80,7 +80,7 @@ where
 
             OrdZSet::<T>::from_tuples((), data)
         } else {
-            OrdZSet::<T>::empty(())
+            OrdZSet::<T>::empty()
         };
         self.time += 1;