… Cargo.lock
… benchmark/run-nexmark.sh
… crates/dbsp/src/operator/dynamic/distinct.rs
… crates/dbsp/src/operator/dynamic/input.rs
… crates/dbsp/src/operator/dynamic/join.rs
… crates/dbsp/src/operator/dynamic/neighborhood.rs
… crates/dbsp/src/operator/dynamic/recursive.rs
… crates/dbsp/src/operator/dynamic/sample.rs
… crates/dbsp/src/storage/buffer_cache/cache.rs
… crates/dbsp/src/storage/file/cache.rs
… crates/dbsp/src/trace/mod.rs
… crates/dbsp/src/trace/ord/fallback/indexed_wset.rs
… crates/dbsp/src/trace/ord/fallback/key_batch.rs
… crates/dbsp/src/trace/ord/fallback/val_batch.rs
… crates/dbsp/src/trace/ord/fallback/wset.rs
… crates/dbsp/src/trace/ord/file/indexed_wset_batch.rs
… crates/dbsp/src/trace/ord/file/wset_batch.rs
… crates/dbsp/src/trace/ord/vec/wset_batch.rs
… crates/dbsp/src/trace/spine_async/merger.rs
… crates/dbsp/src/trace/spine_async/mod.rs
… crates/dbsp/src/trace/spine_async/tests.rs
… crates/dbsp/src/trace/spine_async/thread.rs
… crates/dbsp/src/trace/test/mod.rs
… crates/dbsp/src/typed_batch.rs