This commit rewrites the asynchronous merger to be more asynchronous and faster. It is also simpler--this commit deletes more lines than it adds.

As of now, this is still going through CI, but I thought it was important to get review started.

I'm also running some benchmarks for comparative purposes, but here are a few to get started:

- Without storage: without this series, Nexmark q9 takes 123 s and 114 GB RAM for 100M events; with this series, 56 s and 18 GB RAM
- With storage: without this series, 125 s and 47 GB RAM; with this series, 76 s and 15 GB RAM.

(The RAM above is peak; the asverage would be smaller.)