@blp 2024-07-29
CI passed: https://github.com/feldera/feldera/actions/runs/10137917280