@blp 2024-07-29
> @blp Nice! Looks q9's memory usage has dropped dramatically with this PR.

Yep! Most of them dropped.

> q15 seems to have gotten slightly slower.

q15 is really odd. It is by far the slowest query, even though it's a simple query. I plan to investigate it at some point.

q15 is just:

```
CREATE VIEW q15 AS
SELECT
     FORMAT_DATE('yyyy-MM-dd', date_time) as 'day',
     count(*) AS total_bids,
     count(*) filter (where price < 10000) AS rank1_bids,
     count(*) filter (where price >= 10000 and price < 1000000) AS rank2_bids,
     count(*) filter (where price >= 1000000) AS rank3_bids,
     count(distinct bidder) AS total_bidders,
     count(distinct bidder) filter (where price < 10000) AS rank1_bidders,
     count(distinct bidder) filter (where price >= 10000 and price < 1000000) AS rank2_bidders,
     count(distinct bidder) filter (where price >= 1000000) AS rank3_bidders,
     count(distinct auction) AS total_auctions,
     count(distinct auction) filter (where price < 10000) AS rank1_auctions,
     count(distinct auction) filter (where price >= 10000 and price < 1000000) AS rank2_auctions,
     count(distinct auction) filter (where price >= 1000000) AS rank3_auctions
FROM bid
GROUP BY FORMAT_DATE('yyyy-MM-dd', date_time);
```

> There seems to be this weird thing where we're faster with storage in several cases, but I assume that's measurement variability.

It's fairly consistent in some cases. I haven't investigated, so I can only guess, and your guess is probably as good as mine at this point.