@gz 2024-07-29
The one thing I noticed about q15 which is different from others is that it seems to create a lot more spines