@lalithsuresh 2024-07-29
Probably worth looking at the query plan and profile for oddities (but separately from this PR).