@mihaibudiu 2024-07-29
I will take a look at the plan. But each distinct requires a spine, and each aggregate requires 2.