@mihaibudiu 2024-07-29
This is Q15's plan:
![final](https://github.com/user-attachments/assets/b1b9d346-b2db-4f35-9a7b-8ac270a9b7aa)