@mihaibudiu 2024-07-29
everything red has state, i.e., spine