@mihaibudiu 2024-07-29
> Would multi-join help?

All these joins are joining on the empty key since these are global aggregates. 
I expect a multi-join version of that won't help.