@mihaibudiu 2024-07-29
I mean, there will be fewer spines, but all of them should be tiny