@ryzhyk 2024-07-29
> I don't really see anything that could be improved here

- Shouldn't we be able to use only two distinct's here?
- Can we combine several aggregates into one?
- multiway join would definitely help too