@mihaibudiu 2024-07-29
> > I don't really see anything that could be improved here
> 
>     * Shouldn't we be able to use only two distinct's here?

Each distinct is applied to a different value (a function of the data). I don't see how you can can combine them.

 
>     * Can we combine several aggregates into one?

The aggregate different sets. Moreover, most aggregates are linear. When we improve our aggregate implementations *all* aggregates will look like this.
 
>     * multiway join would definitely help too

Each join has 1 input and 1 output record.