@mihaibudiu 2024-07-29
I expect all memory consumption is in the integrators for the distinct operators.
Only lateness information may help that.
Maybe @blp already runs the query with lateness.