@mihaibudiu 2024-07-29
This is an "infinite state" query, where none of the bidders or auctions can be GC-ed due to the distinct.
So the LATENESS doesn't help at all.
On a side note, we should make the ID fields primary keys if we can (they must be non-null), and we should mark the corresponding fields foreign keys, I plan to add optimizations that use this information.